luffy/luffyapi/requirement.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==3.2.18
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
dnspython==2.4.2
eventlet==0.34.2
greenlet==3.0.3
kombu==5.3.4
Pillow==10.1.0
prompt-toolkit==3.0.43
PyJWT==1.7.1
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
qcloudsms-py==0.1.4
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
